Completion of merge
